unify spike/spike multiple dialogs into a single function
